requirements.txt
PyYAML
pytest
pytest-cov
wheel
twine
